[Issue 63]: Refactor to request utils in result fixing the python 2 unbound issue. (c) Copyright IBM Corp. 2015
