+++ kernel_gateway/tests/services/notebooks/test_request_utils.py
@@ -2,7 +2,7 @@
 # Distributed under the terms of the Modified BSD License.
 
 from kernel_gateway.services.notebooks.request_utils import *
-import unittest
+import unittest, json
 
 class MockRequest(dict):
     def __init__(self, *args, **kwargs):
@@ -107,3 +107,23 @@ class TestRequestUtils(unittest.TestCase):
     def test_headers_to_dict_with_no_headers(self):
         result = headers_to_dict(MockHeaders([]))
         self.assertEqual(result, {},'Empty headers handled incorrectly and did not make empty dict')
+
+        def test_format_request_code_not_escaped(self):
+            '''Test formatting request code without escaped quotes'''
+        test_request = ('{"body": "", "headers": {"Accept-Language": "en-US,en;q=0.8", '
+                        '"If-None-Match": "9a28a9262f954494a8de7442c63d6d0715ce0998", '
+                        '"Accept-Encoding": "gzip, deflate, sdch"}, "args": {}, "path": {}}')
+        request_code = format_request(test_request)
+        #Get the value of REQUEST = "{ to test for equality
+        test_request_js_value = request_code[request_code.index("\"{"):]
+        self.assertEqual(test_request, json.loads(test_request_js_value), "Request code without escaped quotes was not formatted correctly")
+
+    def test_format_request_code_escaped(self):
+        '''Test formatting request code where multiple escaped quotes exist'''
+        test_request = ('{"body": "", "headers": {"Accept-Language": "en-US,en;q=0.8", '
+                        '"If-None-Match": "\"\"9a28a9262f954494a8de7442c63d6d0715ce0998\"\"", '
+                        '"Accept-Encoding": "gzip, deflate, sdch"}, "args": {}, "path": {}}')
+        request_code = format_request(test_request)
+        #Get the value of REQUEST = "{ to test for equality
+        test_request_js_value = request_code[request_code.index("\"{"):]
+        self.assertEqual(test_request, json.loads(test_request_js_value), "Escaped Request code was not formatted correctly")
